rename stream-object.ts -> output-format-handlers.ts
