+++ packages/core/src/stream/base/output-format-handlers.ts
